support optional unified KML file
