+++ frontend/react/src/queries/PermitsQuery.js
@@ -60,6 +60,7 @@ const PERMITS_QUERY = gql`
           issueDate
           notes
           imageUrl
+          kmlUrl
           name
           bounds {
             geojson
